Move oci package to pkg
