+++ Dockerfile
@@ -6,6 +6,7 @@ COPY go.sum go.sum
 RUN go mod download
 COPY main.go main.go
 COPY internal/ internal/
+COPY pkg/ pkg/
 RUN CGO_ENABLED=0 go build -installsuffix 'static' -o spegel .
 
 FROM gcr.io/distroless/static:nonroot
