+++ internal/registry/distribution_test.go
@@ -1,4 +1,4 @@
-package oci
+package registry
 
 import (
 	"testing"
@@ -14,7 +14,7 @@ func TestParsePathComponents(t *testing.T) {
 		path            string
 		expectedRef     string
 		expectedDgst    digest.Digest
-		expectedRefType ReferenceType
+		expectedRefType referenceType
 	}{
 		{
 			name:            "valid manifest tag",
@@ -22,7 +22,7 @@ func TestParsePathComponents(t *testing.T) {
 			path:            "/v2/foo/bar/manifests/hello-world",
 			expectedRef:     "example.com/foo/bar:hello-world",
 			expectedDgst:    "",
-			expectedRefType: ReferenceTypeManifest,
+			expectedRefType: referenceTypeManifest,
 		},
 		{
 			name:            "valid blob digest",
@@ -30,12 +30,12 @@ func TestParsePathComponents(t *testing.T) {
 			path:            "/v2/library/nginx/blobs/sha256:295c7be079025306c4f1d65997fcf7adb411c88f139ad1d34b537164aa060369",
 			expectedRef:     "",
 			expectedDgst:    digest.Digest("sha256:295c7be079025306c4f1d65997fcf7adb411c88f139ad1d34b537164aa060369"),
-			expectedRefType: ReferenceTypeBlob,
+			expectedRefType: referenceTypeBlob,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			ref, dgst, refType, err := ParsePathComponents(tt.registry, tt.path)
+			ref, dgst, refType, err := parsePathComponents(tt.registry, tt.path)
 			require.NoError(t, err)
 			require.Equal(t, tt.expectedRef, ref)
 			require.Equal(t, tt.expectedDgst, dgst)
@@ -45,11 +45,11 @@ func TestParsePathComponents(t *testing.T) {
 }
 
 func TestParsePathComponentsInvalidPath(t *testing.T) {
-	_, _, _, err := ParsePathComponents("example.com", "/v2/xenitab/spegel/v0.0.1")
+	_, _, _, err := parsePathComponents("example.com", "/v2/xenitab/spegel/v0.0.1")
 	require.EqualError(t, err, "distribution path could not be parsed")
 }
 
 func TestParsePathComponentsMissingRegistry(t *testing.T) {
-	_, _, _, err := ParsePathComponents("", "/v2/xenitab/spegel/manifests/v0.0.1")
+	_, _, _, err := parsePathComponents("", "/v2/xenitab/spegel/manifests/v0.0.1")
 	require.EqualError(t, err, "registry parameter needs to be set for tag references")
 }
